Refacrored reducers
