+++ src/store/reducers/app-data/app-data.js
@@ -1,26 +1,15 @@
 import {ActionType} from "../../action";
 
-import {ALL_GENRES} from "../../../const";
 import {extend} from "../../../utils";
 
 const initialState = {
-  activeGenre: ALL_GENRES,
   movies: [],
-  moviesByGenre: [],
   promo: {},
-  shownMovies: 0,
+  reviews: [],
 };
 
-export const movies = (state = initialState, action) => {
+export const appData = (state = initialState, action) => {
   switch (action.type) {
-    case ActionType.GET_MOVIES_BY_GENRE:
-      return extend(state, {
-        activeGenre: action.payload,
-      });
-    case ActionType.SHOW_MORE_MOVIES:
-      return extend(state, {
-        shownMovies: action.payload,
-      });
     case ActionType.GET_ALL_MOVIES:
       return extend(state, {
         movies: action.payload,
@@ -28,6 +17,10 @@ export const movies = (state = initialState, action) => {
     case ActionType.GET_PROMO_MOVIE:
       return extend(state, {
         promo: action.payload,
+      });
+    case ActionType.GET_MOVIE_REVIEWS:
+      return extend(state, {
+        reviews: action.payload,
       });
   }
